% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent_multiple.R
\name{gradient_descent_multiple}
\alias{gradient_descent_multiple}
\title{Gradient descent for multiple linear regression}
\usage{
gradient_descent_multiple(
  X,
  y,
  learning_rate = 0.01,
  max_itr = 1000,
  tol = 1e-05
)
}
\arguments{
\item{X}{the predictors (an n*p matrix). X should be standardized first.}

\item{y}{the response variable (a vector)}

\item{learning_rate}{the step size for gradient descent}

\item{max_itr}{the maximum number of iterations to run gradient descent}

\item{tol}{the tolearnace between current and best cost}
}
\value{
Theta estimated parameters

cost cost history

best_cost best cost

itr number of iterations conducted
}
\description{
Perform gradient descent for linear regression with multiple predictors
}
\examples{
# Generate X with n=100, p=5
X <- matrix(rnorm(100*5, mean = 0, sd = 1), 100, 5)
Theta = c(1, 0.5, -5, 2, 3)
y <- X\%*\%Theta + rnorm(100, mean = 0, sd = 1)
# Perform gradient descent
result <- gradient_descent_multiple(X, y, learning_rate = 0.01, max_itr = 10000, tol = 1e-5)
Thetahat = result$Theta
result$itr
result$best_cost
cost = result$cost
# plot the cost history over iterations
plot(1:length(cost), cost, xlab = "iterations", main = "Cost over iterations")
}
