% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_descent_single.R
\name{gradient_descent_single}
\alias{gradient_descent_single}
\title{Gradient descent for simple linear regression}
\usage{
gradient_descent_single(
  X,
  y,
  learning_rate = 0.01,
  max_itr = 1000,
  tol = 1e-05
)
}
\arguments{
\item{X}{the predictor (a p-dimensional vector).}

\item{y}{the response variable (a vector)}

\item{learning_rate}{the step size for gradient descent}

\item{max_itr}{the maximum number of iterations to run gradient descent}

\item{tol}{the tolearnace between current and best cost}
}
\value{
Theta estimated parameters

cost cost history

best_cost best cost

itr number of iterations conducted
}
\description{
Perform gradient descent for linear regression with one predictor and intercept
}
\examples{
# Generate some random data
X <- rnorm(100, mean = 0, sd = 1)
y <- 5 + 2 * X + rnorm(100, mean = 0, sd = 1)
# Perform gradient descent
result <- gradient_descent_single(X, y, learning_rate = 0.01, max_itr = 10000, tol = 1e-5)
# Print the result
print(paste("Intercept:", result$intercept))
print(paste("Slope:", result$slope))
}
